requirements.txt
anthropic>=0.25.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
flask>=3.0.0
openpyxl>=3.1.0
markdown>=3.5.0
